mob-typer fixes to use common logic from recon
