+++ mob_suite/mob_typer.py
@@ -6,39 +6,34 @@ from argparse import (ArgumentParser, FileType)
 from ete3 import NCBITaxa
 from mob_suite.version import __version__
 import mob_suite.mob_init
+from collections import OrderedDict
+from operator import itemgetter
+from mob_suite.blast import BlastReader,BlastRunner
 from mob_suite.wrappers import mash
-from mob_suite.utils import \
-    replicon_blast, \
-    mob_blast, \
-    getRepliconContigs, \
-    fix_fasta_header, \
-    getMashBestHit, \
+from mob_suite.utils import fix_fasta_header, \
     calcFastaStats, \
     verify_init, \
     check_dependencies, \
     read_sequence_info, \
-    read_file_to_dict,\
-    dict_from_alt_key_list, \
-    get_data_associated_with_key, \
     writeReport, \
-    getMashBestHitMultiSeq, \
-    calcFastaStatsIndividual, \
     sort_biomarkers, \
-    isETE3DBTAXAFILEexists, \
-    initETE3Database, \
-    ETE3_db_status_check
+    ETE3_db_status_check, \
+    calc_md5, \
+    GC, \
+    read_fasta_dict, \
+    identify_biomarkers, \
+    parseMash, \
+    determine_mpf_type, \
+    hostrange, \
+    dict_from_alt_key_list, \
+    read_file_to_dict
 
 
 from mob_suite.constants import  ETE3DBTAXAFILE, \
-    LIT_PLASMID_TAXONOMY_HEADER, \
     MOB_TYPER_REPORT_HEADER, \
     MOB_CLUSTER_INFO_HEADER, \
-    MOB_RECON_INFO_HEADER,\
     default_database_dir, \
-    NCBI_PLASMID_TAXONOMY_FILE, \
-    NCBI_PLASMID_TAXONOMY_HEADER, \
-    LIT_PLASMID_TAXONOMY_FILE, \
-    ETE3_LOCK_FILE
+    ETE3_LOCK_FILE, NCBI_PLASMID_TAXONOMY_FILE, NCBI_PLASMID_TAXONOMY_HEADER, LIT_PLASMID_TAXONOMY_FILE, LIT_PLASMID_TAXONOMY_HEADER
 
 def init_console_logger(lvl=2):
     root = logging.getLogger()
@@ -52,51 +47,45 @@ def init_console_logger(lvl=2):
     logging.basicConfig(format=LOG_FORMAT, level=report_lvl)
     return logging.getLogger(__name__)
 
-
 def parse_args():
     "Parse the input arguments, use '-h' for help"
-    default_database_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'databases')
-    parser = ArgumentParser(
-        description="MOB-Typer: Plasmid typing version: {}".format(
-            __version__))
 
+    parser = ArgumentParser(
+        description="MOB-Typer: Plasmid typing and mobility prediction: {}".format(
+            __version__))
+    parser.add_argument('-i', '--infile', type=str, required=True, help='Input assembly fasta file to process')
     parser.add_argument('-o', '--out_file', type=str, required=True, help='Output file to write results')
     parser.add_argument('-a', '--analysis_dir', type=str, required=False, help='Analysis directory')
-    parser.add_argument('-i', '--infile', type=str, required=True, help='Input assembly fasta file to process')
-    parser.add_argument('-s', '--sample_id', type=str, required=False, help='Sample name for report')
-    parser.add_argument('-x', '--multi', required=False, help='Fasta file is multiple independant records',action='store_true',
-                        default=False)
-
     parser.add_argument('-n', '--num_threads', type=int, required=False, help='Number of threads to be used', default=1)
+    parser.add_argument('-s', '--sample_id', type=str, required=False, help='Sample Prefix for reports')
+    parser.add_argument('-f', '--force', required=False, help='Overwrite existing directory',
+                        action='store_true')
+    parser.add_argument('-x', '--multi', required=False, help='Treat each sequence as an independant plasmid',
+                        action='store_true')
 
     parser.add_argument('--min_rep_evalue', type=str, required=False,
                         help='Minimum evalue threshold for replicon blastn',
                         default=0.00001)
-
     parser.add_argument('--min_mob_evalue', type=str, required=False,
                         help='Minimum evalue threshold for relaxase tblastn',
                         default=0.00001)
-
     parser.add_argument('--min_con_evalue', type=str, required=False, help='Minimum evalue threshold for contig blastn',
                         default=0.00001)
+    parser.add_argument('--min_rpp_evalue', type=str, required=False,
+                        help='Minimum evalue threshold for repetitve elements blastn',
+                        default=0.00001)
 
-    parser.add_argument('--min_ori_evalue', type=str, required=False,
-                        help='Minimum evalue threshold for oriT elements blastn',
-                        default=0.00001)
-    parser.add_argument('--min_mpf_evalue', type=str, required=False,
-                        help='Minimum evalue threshold for mpf elements blastn',
-                        default=0.00001)
+    parser.add_argument('--min_length', type=str, required=False, help='Minimum length of contigs to classify',
+                        default=1000)
 
     parser.add_argument('--min_rep_ident', type=int, required=False, help='Minimum sequence identity for replicons',
                         default=80)
-
     parser.add_argument('--min_mob_ident', type=int, required=False, help='Minimum sequence identity for relaxases',
                         default=80)
-
-    parser.add_argument('--min_ori_ident', type=int, required=False,
-                        help='Minimum sequence identity for oriT elements', default=90)
-    parser.add_argument('--min_mpf_ident', type=int, required=False,
-                        help='Minimum sequence identity for mpf elements', default=80)
+    parser.add_argument('--min_con_ident', type=int, required=False, help='Minimum sequence identity for contigs',
+                        default=80)
+    parser.add_argument('--min_rpp_ident', type=int, required=False,
+                        help='Minimum sequence identity for repetitive elements', default=80)
 
     parser.add_argument('--min_rep_cov', type=int, required=False,
                         help='Minimum percentage coverage of replicon query by input assembly',
@@ -106,297 +95,66 @@ def parse_args():
                         help='Minimum percentage coverage of relaxase query by input assembly',
                         default=80)
 
-    parser.add_argument('--min_ori_cov', type=int, required=False,
-                        help='Minimum percentage coverage of oriT',
-                        default=90)
-    parser.add_argument('--min_mpf_cov', type=int, required=False,
-                        help='Minimum percentage coverage of mpf',
+    parser.add_argument('--min_con_cov', type=int, required=False,
+                        help='Minimum percentage coverage of assembly contig by the plasmid reference database to be considered',
+                        default=70)
+
+    parser.add_argument('--min_rpp_cov', type=int, required=False,
+                        help='Minimum percentage coverage of contigs by repetitive elements',
                         default=80)
 
     parser.add_argument('--min_overlap', type=int, required=False,
                         help='Minimum overlap of fragments',
                         default=10)
 
-    parser.add_argument('--keep_tmp', required=False,help='Do not delete temporary file directory', action='store_true')
+    parser.add_argument('-k', '--keep_tmp', required=False, help='Do not delete temporary file directory',
+                        action='store_true')
+
     parser.add_argument('--debug', required=False, help='Show debug information', action='store_true')
-    parser.add_argument('-m','--plasmid_meta', type=str, required=False,
-                        help='MOB-cluster plasmid cluster formatted file matched to the reference plasmid db',
-                        default=os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                             'databases/clusters.txt'))
+
     parser.add_argument('--plasmid_mash_db', type=str, required=False,
                         help='Companion Mash database of reference database',
-                        default=os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                             'databases/ncbi_plasmid_full_seqs.fas.msh'))
+                        default=os.path.join(default_database_dir,
+                                             'ncbi_plasmid_full_seqs.fas.msh'))
+    parser.add_argument('-m','--plasmid_meta', type=str, required=False,
+                        help='MOB-cluster plasmid cluster formatted file matched to the reference plasmid db',
+                        default=os.path.join(default_database_dir,
+                                             'clusters.txt'))
+    parser.add_argument('--plasmid_db_type', type=str, required=False, help='Blast database type of reference database',
+                        default='blastn')
     parser.add_argument('--plasmid_replicons', type=str, required=False, help='Fasta of plasmid replicons',
-                        default=os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                             'databases/rep.dna.fas'))
+                        default=os.path.join(default_database_dir,
+                                             'rep.dna.fas'))
+    parser.add_argument('--repetitive_mask', type=str, required=False, help='Fasta of known repetitive elements',
+                        default=os.path.join(default_database_dir,
+                                             'repetitive.dna.fas'))
     parser.add_argument('--plasmid_mob', type=str, required=False, help='Fasta of plasmid relaxases',
-                        default=os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                             'databases/mob.proteins.faa'))
+                        default=os.path.join(default_database_dir,
+                                             'mob.proteins.faa'))
     parser.add_argument('--plasmid_mpf', type=str, required=False, help='Fasta of known plasmid mate-pair proteins',
-                        default=os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                             'databases/mpf.proteins.faa'))
+                        default=os.path.join(default_database_dir,
+                                             'mpf.proteins.faa'))
     parser.add_argument('--plasmid_orit', type=str, required=False, help='Fasta of known plasmid oriT dna sequences',
-                        default=os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                                             'databases/orit.fas'))
+                        default=os.path.join(default_database_dir,
+                                             'orit.fas'))
+    parser.add_argument('-d', '--database_directory',
+                        default=default_database_dir,
+                        required=False,
+                        help='Directory you want to use for your databases. If the databases are not already '
+                             'downloaded, they will be downloaded automatically. Defaults to {}'.format(default_database_dir))
     parser.add_argument('--primary_cluster_dist', type=int, required=False, help='Mash distance for assigning primary cluster id 0 - 1', default=0.06)
     parser.add_argument('--secondary_cluster_dist', type=int, required=False, help='Mash distance for assigning primary cluster id 0 - 1',
                         default=0.025)
-    parser.add_argument('-d','--database_directory',default=default_database_dir,
-                        required=False,
-                        help='Directory you want to use for your databases. If the databases are not already '
-                             'downloaded, they will be downloaded automatically. Defaults to {}. '
-                             'If you change this from the default, will override --plasmid_mash_db, '
-                             '--plasmid_replicons, --plasmid_mob, --plasmid_mpf, and '
-                             '--plasmid_orit'.format(default_database_dir))
+    parser.add_argument('-V', '--version', action='version', version="%(prog)s " + __version__ )
 
-    parser.add_argument('-V', '--version', action='version', version="%(prog)s " + __version__)
+
     return parser.parse_args()
-
-def determine_mpf_type(hits):
-    types = dict()
-    for hit in hits:
-        if not hit in types:
-            types[hit] = 0
-        types[hit] += 1
-
-    return max(types, key=lambda i: types[i])
-
-def getAssocValues(query_list_values,look_up_key,value_key,data):
-    values = []
-
-    for q in query_list_values:
-        values.extend(get_data_associated_with_key(look_up_key, q, value_key, data))
-    return  values
 
 def initMOBTyperReportTemplate(header):
     data = {}
     for i in header:
         data[i] = '-'
     return data
-
-def getHeirarchy(taxid):
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.info("Did not find taxa.sqlite in {}. Initializaing ete3 taxonomy database".format(ETE3DBTAXAFILE))
-        initETE3Database()
-
-    ncbi = NCBITaxa(dbfile=ETE3DBTAXAFILE)
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.error("Tried ete3 init, but still was not able to find taxa.sqlite file for ete3 lib in {}. Aborting".format(ETE3DBTAXAFILE))
-        return ['-','-']
-
-    if not isinstance(taxid,int):
-        return {'names': [], 'ranks': []}
-
-    lineage = ncbi.get_lineage(taxid)
-    names = ncbi.get_taxid_translator(lineage)
-    ranks = []
-    for id in lineage:
-        ranks.append(ncbi.get_rank(id))
-
-    return {'names': names, 'ranks': names}
-
-
-
-
-def filter_invalid_taxids(taxids):
-    filtered = []
-    for i in taxids:
-        if i == 'NaN' or i == 'nan' or i == None or i == '-' :
-            continue
-
-        try:
-            filtered.append(int(i))
-
-        except ValueError:
-            continue
-
-    return filtered
-
-
-def getHeirarchy(taxid):
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.info("Did not find taxa.sqlite in {}. Initializaing ete3 taxonomy database".format(ETE3DBTAXAFILE))
-        initETE3Database()
-
-    ncbi = NCBITaxa(dbfile=ETE3DBTAXAFILE)
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.error("Tried ete3 init, but still was not able to find taxa.sqlite file for ete3 lib in {}. Aborting".format(ETE3DBTAXAFILE))
-        return ['-','-']
-
-    if not isinstance(taxid,int):
-        return {'names': [], 'ranks': []}
-
-    lineage = ncbi.get_lineage(taxid)
-    names = ncbi.get_taxid_translator(lineage)
-    ranks = []
-    for id in lineage:
-        ranks.append(ncbi.get_rank(id))
-
-    return {'names': names, 'ranks': names}
-
-
-def getTaxid(taxon):
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.info("Did not find taxa.sqlite in {}. Initializaing ete3 taxonomy database".format(ETE3DBTAXAFILE))
-        initETE3Database()
-
-    ncbi = NCBITaxa(dbfile=ETE3DBTAXAFILE)
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.error("Tried ete3 init, but still was not able to find taxa.sqlite file for ete3 lib in {}. Aborting".format(ETE3DBTAXAFILE))
-        return ['-','-']
-
-    taxid = ncbi.get_name_translator(taxon)
-
-    return  taxid
-
-
-def getTaxonConvergence(taxids):
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.info("Did not find taxa.sqlite in {}. Initializaing ete3 taxonomy database".format(ETE3DBTAXAFILE))
-        initETE3Database(ETE3DBTAXAFILE)
-
-    ncbi = NCBITaxa(dbfile=ETE3DBTAXAFILE)
-
-    if not isETE3DBTAXAFILEexists(ETE3DBTAXAFILE):
-        logging.error(
-            "Tried ete3 init, but still was not able to find taxa.sqlite file for ete3 lib in {}. Aborting".format(
-                ETE3DBTAXAFILE))
-        return ['-', '-']
-
-    tax_ranks = ['genus','family','order','class','phylum',  'species']
-
-    phylla = []
-    lineages = []
-    if len(taxids) == 0:
-        return ['-', '-']
-
-    elif len(taxids) == 1:
-        common_taxids = ncbi.get_lineage(taxids[0])
-
-    else:
-        for taxid in taxids:
-            lineage = ncbi.get_lineage(taxid)
-            ranks = ncbi.get_rank(lineage)
-            names = ncbi.get_taxid_translator(lineage)
-            for id in ranks:
-                rank = ranks[id]
-                name = names[id]
-                if rank == 'phylum':
-                    phylla.append(name)
-
-            lineages.append(lineage)
-
-        common_taxids = lineages[0]
-        for i in range(1,len(lineages)):
-            common_taxids = list(set(common_taxids) - set(set(common_taxids) - set(lineages[i])))
-
-
-
-    if len(common_taxids) == 0:
-        return ['-', '-']
-
-    taxa = {}
-
-    for i in range(0,len(common_taxids)):
-        rank = ncbi.get_rank([common_taxids[i]])[common_taxids[i]]
-        name = ncbi.get_taxid_translator([common_taxids[i]])[common_taxids[i]]
-
-        if not rank in tax_ranks:
-            continue
-
-        taxa[rank] = name
-
-    for rank in tax_ranks:
-        if rank in taxa:
-            return [rank,taxa[rank]]
-
-    phylla = list(set(phylla))
-    sorted(phylla)
-    if len(phylla) > 1:
-        return ('multi-phylla',",".join(phylla))
-    elif len(phylla) == 1:
-        return ('phylum', ",".join(phylla))
-
-    return(['-','-'])
-
-
-
-
-def hostrange(replion_types, relaxase_types, mob_cluster_id,ncbi,lit):
-
-    host_range_predictions = {
-        'observed_host_range_ncbi_name':'',
-        'observed_host_range_ncbi_rank':'',
-        'reported_host_range_lit_name':'',
-        'reported_host_range_lit_rank':'',
-        'predicted_host_range_overall_name':'',
-        'predicted_host_range_overall_rank':'',
-        'associated_pmid(s)':'-'
-
-    }
-
-
-    #Determine taxids associated with NCBI records
-    if mob_cluster_id != '-' :
-        ncbi_cluster_taxids = getAssocValues([mob_cluster_id], 'primary_cluster_id', 'taxid', ncbi)
-
-    else:
-        ncbi_cluster_taxids = []
-
-    if len(replion_types) > 0 :
-        ncbi_replicon_taxids = getAssocValues(replion_types,'rep_type(s)','taxid',ncbi)
-        lit_replicon_taxids = getAssocValues(replion_types, 'rep_type(s)', 'reported_host_range_taxid', lit)
-        lit_replicon_taxids = lit_replicon_taxids + getAssocValues(replion_types, 'rep_type(s)', 'host_taxid', lit)
-        pmids = list(set(getAssocValues(replion_types, 'rep_type(s)', 'pmid', lit)))
-        sorted(pmids)
-        pmids = filter_invalid_taxids(list(set(pmids)))
-        host_range_predictions['associated_pmid(s)'] = '; '.join(str(x) for x in pmids)
-
-    else:
-        ncbi_replicon_taxids = []
-        lit_replicon_taxids = []
-
-    if len(relaxase_types) > 0:
-        ncbi_relaxase_taxids = getAssocValues(relaxase_types, 'relaxase_type_accession(s)', 'taxid', ncbi)
-    else:
-        ncbi_relaxase_taxids = []
-
-    if  len(ncbi_cluster_taxids) == 0 and len(ncbi_replicon_taxids) == 0 and len(ncbi_relaxase_taxids) == 0:
-        host_range_predictions['observed_host_range_ncbi_rank'] = '-'
-        host_range_predictions['observed_host_range_ncbi_name'] = '-'
-        ncbi_unique_taxids = []
-    else:
-        ncbi_unique_taxids = filter_invalid_taxids(list(set(ncbi_replicon_taxids + ncbi_cluster_taxids + ncbi_relaxase_taxids)))
-        host_range_predictions['observed_host_range_ncbi_rank'], host_range_predictions['observed_host_range_ncbi_name'] = getTaxonConvergence(ncbi_unique_taxids)
-
-
-    # Determine taxids associated with literature
-
-    lit_unique_taxids = filter_invalid_taxids(list(set(lit_replicon_taxids )))
-
-    host_range_predictions['reported_host_range_lit_rank'], host_range_predictions['reported_host_range_lit_name'] = getTaxonConvergence(lit_unique_taxids)
-
-    #determine overall host range
-    overall_taxids = filter_invalid_taxids(list(set(ncbi_unique_taxids + lit_unique_taxids)))
-    host_range_predictions['predicted_host_range_overall_rank'], host_range_predictions['predicted_host_range_overall_name'] = getTaxonConvergence(overall_taxids)
-
-    #move host-range prediction up to family when it is at genus or species level
-    if host_range_predictions['predicted_host_range_overall_rank'] == 'genus' or host_range_predictions['predicted_host_range_overall_rank'] == 'species':
-        taxid = getTaxid(host_range_predictions['predicted_host_range_overall_name'])
-        heir = getHeirarchy(taxid)
-        names = heir['names']
-        ranks = heir['ranks']
-
-        for i in range(0,len(ranks)):
-            if ranks[i] == 'Family':
-                host_range_predictions['predicted_host_range_overall_rank'] = 'family'
-                host_range_predictions['predicted_host_range_overall_name'] = names[i]
-
-
-    return host_range_predictions
-
-
 
 
 def getBioMarkerContigs(contig_dict):
@@ -521,18 +279,20 @@ def main():
 
     if database_dir == default_database_dir:
         mob_ref = args.plasmid_mob
-        mpf_ref = args.plasmid_mpf
-        orit_ref = args.plasmid_orit
         mash_db = args.plasmid_mash_db
         replicon_ref = args.plasmid_replicons
         plasmid_meta = args.plasmid_meta
+        repetitive_mask_file = args.repetitive_mask
+        mpf_ref = args.plasmid_mpf
+        plasmid_orit = args.plasmid_orit
     else:
         mob_ref = os.path.join(database_dir, 'mob.proteins.faa')
-        mpf_ref = os.path.join(database_dir, 'mpf.proteins.faa')
-        orit_ref = os.path.join(database_dir, 'orit.fas')
         mash_db = os.path.join(database_dir, 'ncbi_plasmid_full_seqs.fas.msh')
         replicon_ref = os.path.join(database_dir, 'rep.dna.fas')
         plasmid_meta = os.path.join(database_dir, 'clusters.txt')
+        repetitive_mask_file = os.path.join(database_dir, 'repetitive.dna.fas')
+        mpf_ref = os.path.join(database_dir, 'mpf.proteins.faa')
+        plasmid_orit = os.path.join(database_dir, 'orit.fas')
 
 
     fixed_fasta = os.path.join(tmp_dir, 'fixed.input.fasta')
@@ -540,6 +300,8 @@ def main():
     mob_blast_results = os.path.join(tmp_dir, 'mobtyper_blast_results.txt')
     mpf_blast_results = os.path.join(tmp_dir, 'mpf_blast_results.txt')
     orit_blast_results = os.path.join(tmp_dir, 'orit_blast_results.txt')
+    repetitive_blast_results = os.path.join(tmp_dir, 'repetitive_blast_results.txt')
+    mash_file = os.path.join(tmp_dir, "mash_{}.txt".format(sample_id))
 
     if os.path.isfile(mob_blast_results):
         os.remove(mob_blast_results)
@@ -551,15 +313,13 @@ def main():
         os.remove(replicon_blast_results)
 
 
-    mash_file = os.path.join(tmp_dir, "mash_{}.txt".format(sample_id))
-
-
     # Input numeric params
 
     min_rep_ident = float(args.min_rep_ident)
     min_mob_ident = float(args.min_mob_ident)
-    min_ori_ident = float(args.min_ori_ident)
-    min_mpf_ident = float(args.min_mpf_ident)
+    min_ori_ident = float(args.min_rep_ident)
+    min_mpf_ident = float(args.min_mob_ident)
+    min_rpp_ident = float(args.min_rpp_ident)
 
     idents = {'min_rep_ident': min_rep_ident, 'min_mob_ident': min_mob_ident, 'min_ori_ident': min_ori_ident}
 
@@ -577,8 +337,9 @@ def main():
 
     min_rep_cov = float(args.min_rep_cov)
     min_mob_cov = float(args.min_mob_cov)
-    min_ori_cov = float(args.min_ori_cov)
-    min_mpf_cov = float(args.min_mpf_cov)
+    min_ori_cov = float(args.min_rep_cov)
+    min_mpf_cov = float(args.min_mob_cov)
+    min_rpp_cov = float(args.min_rpp_cov)
 
 
 
@@ -599,8 +360,9 @@ def main():
 
     min_rep_evalue = float(args.min_rep_evalue)
     min_mob_evalue = float(args.min_mob_evalue)
-    min_ori_evalue = float(args.min_ori_evalue)
-    min_mpf_evalue = float(args.min_mpf_evalue)
+    min_ori_evalue = float(args.min_rep_evalue)
+    min_mpf_evalue = float(args.min_mob_evalue)
+    min_rpp_evalue = float(args.min_mob_evalue)
 
 
     evalues = {'min_rep_evalue': min_rep_evalue, 'min_mob_evalue': min_mob_evalue, 'min_con_evalue': min_ori_evalue}
@@ -627,25 +389,6 @@ def main():
     if not os.path.isdir(tmp_dir):
         os.mkdir(tmp_dir, 0o755)
 
-    mobtyper_results = {}
-    fix_fasta_header(input_fasta, fixed_fasta)
-    if multi:
-        fastaSeqStats = calcFastaStatsIndividual(fixed_fasta)
-        for sid in fastaSeqStats:
-            if not sid in mobtyper_results:
-                mobtyper_results[sid] = initMOBTyperReportTemplate(MOB_TYPER_REPORT_HEADER)
-                mobtyper_results[sid]['sample_id'] = sid
-                mobtyper_results[sid]['num_contigs'] = 1
-                for stat in fastaSeqStats[sid]:
-                    mobtyper_results[sid][stat] = fastaSeqStats[sid][stat]
-    else:
-        mobtyper_results[sample_id] = initMOBTyperReportTemplate(MOB_TYPER_REPORT_HEADER)
-        mobtyper_results[sample_id]['sample_id'] = sample_id
-        fastaSeqStats = calcFastaStats(fixed_fasta)
-        mobtyper_results[sample_id]['md5'] = fastaSeqStats['md5']
-        mobtyper_results[sample_id]['total_length'] = fastaSeqStats['size']
-        mobtyper_results[sample_id]['num_contigs'] = fastaSeqStats['num_seq']
-        mobtyper_results[sample_id]['gc'] = fastaSeqStats['gc_content']
 
     #Test that ETE3 db is ok and lock process check
     dbstatus = ETE3_db_status_check(1, ETE3_LOCK_FILE, ETE3DBTAXAFILE, logging)
@@ -657,174 +400,181 @@ def main():
     reference_sequence_meta = read_sequence_info(plasmid_meta,MOB_CLUSTER_INFO_HEADER)
 
 
+
+    #initilize master record tracking
+    fix_fasta_header(input_fasta, fixed_fasta)
+    contig_seqs = read_fasta_dict(fixed_fasta)
+    contig_info = {}
+    for id in contig_seqs:
+        seq = contig_seqs[id]
+        contig_info[id] = {}
+        for feature in MOB_TYPER_REPORT_HEADER:
+            contig_info[id][feature] = ''
+        contig_info[id]['md5'] = calc_md5(seq)
+        contig_info[id]['gc'] = GC(seq)
+        contig_info[id]['size'] = len(seq)
+        contig_info[id]['contig_id'] = id
+        contig_info[id]['sample_id'] = sample_id
+
+
+    #Makeblastdb
+    blast_runner = BlastRunner(fixed_fasta, tmp_dir)
+    build_success = blast_runner.makeblastdb(fixed_fasta, 'nucl',logging=logging)
+    if build_success == False:
+        logging.error("Could not build blast database, check error messages..cannot continue")
+        sys.exit()
+
     #run individual marker blasts
-    logger.info('Running replicon blast on {}'.format(replicon_ref))
-    replicon_contigs = getRepliconContigs(
-        replicon_blast(replicon_ref, fixed_fasta, min_rep_ident, min_rep_cov, min_rep_evalue, tmp_dir, replicon_blast_results,logging=logging,
-                       num_threads=num_threads))
+    # blast replicon database
 
-    found_replicons = sort_biomarkers(getBioMarkerContigs(replicon_contigs))
-    del(replicon_contigs)
+    #Blast reference databases
 
-    mob_contigs = getRepliconContigs(
-        mob_blast(mob_ref, fixed_fasta, min_mob_ident, min_mob_cov, min_mob_evalue, tmp_dir, mob_blast_results, logging=logging,num_threads=num_threads))
+    contig_info = identify_biomarkers(contig_info, fixed_fasta, tmp_dir, 25, logging, \
+                        replicon_ref, min_rep_ident, min_rep_cov, min_rep_evalue, replicon_blast_results, \
+                        mob_ref, min_mob_ident, min_mob_cov, min_mob_evalue, mob_blast_results, \
+                        mpf_ref, min_mpf_ident, min_mpf_cov, min_mpf_evalue, mpf_blast_results, \
+                        repetitive_mask_file, min_rpp_ident, min_rpp_cov, min_rpp_evalue, \
+                        plasmid_orit, orit_blast_results, repetitive_blast_results, \
+                        num_threads=1)
 
-    found_mob = sort_biomarkers(getBioMarkerContigs(mob_contigs))
-    del(mob_contigs)
 
-    logger.info('Running mpf blast on {}'.format(mob_ref))
-    mpf_contigs = getRepliconContigs(
-        mob_blast(mpf_ref, fixed_fasta, min_mpf_ident, min_mpf_cov, min_mpf_evalue, tmp_dir, mpf_blast_results, logging=logging,num_threads=num_threads))
-
-    found_mpf = sort_biomarkers(getBioMarkerContigs(mpf_contigs))
-    del(mpf_contigs)
-
-    logger.info('Running orit blast on {}'.format(orit_ref))
-    orit_contigs = getRepliconContigs(
-        replicon_blast(orit_ref, fixed_fasta, min_ori_ident, min_ori_cov, min_ori_evalue, tmp_dir, orit_blast_results,logging=logging,
-                       num_threads=num_threads))
-    found_orit = sort_biomarkers(getBioMarkerContigs(orit_contigs))
-    del(orit_contigs)
-
-    # Get closest neighbor by mash distance in the entire plasmid database
 
     m = mash()
-    mashfile_handle = open(mash_file, 'w')
-    mash_top_hit = {}
+    mobtyper_results = []
 
-    if multi:
-        tmp_mash_db_file = os.path.join(tmp_dir,"tmp_mash.msh")
-        m.mashsketch(fixed_fasta, tmp_mash_db_file , num_threads=num_threads)
-        m.run_mash(mash_db, tmp_mash_db_file , mashfile_handle)
-        mash_results = m.read_mash(mash_file)
-        mash_top_hit = getMashBestHitMultiSeq(mash_results)
-
-    else:
-
-        m.run_mash(mash_db, fixed_fasta, mashfile_handle)
-        mash_results = m.read_mash(mash_file)
-        mash_top_hit[sample_id] = getMashBestHit(mash_results)
-
-
-    #Get MOB-cluster information for reference database matches
-    mob_cluster_memberships = {}
-
-    for s in mash_top_hit:
-
-        if not s in mob_cluster_memberships:
-            mob_cluster_memberships[s] = {'primary_cluster_id':'-','secondary_cluster_id':'-'}
-
-        plasmid_primary_acs = '-'
-        plasmid_secondary_acs = '-'
-
-        if mash_top_hit[s]['top_hit'] in reference_sequence_meta:
-            if mash_top_hit[s]['mash_hit_score'] <= primary_distance:
-                plasmid_primary_acs = reference_sequence_meta[mash_top_hit[s]['top_hit']]['primary_cluster_id']
-            if mash_top_hit[s]['mash_hit_score'] <= secondary_distance:
-                plasmid_secondary_acs = reference_sequence_meta[mash_top_hit[s]['top_hit']]['secondary_cluster_id']
-
-            mob_cluster_memberships[s]['primary_cluster_id'] = plasmid_primary_acs
-            mob_cluster_memberships[s]['secondary_cluster_id'] = plasmid_secondary_acs
-            mob_cluster_memberships[s]['mash_nearest_neighbor'] = mash_top_hit[s]['top_hit']
-            mob_cluster_memberships[s]['mash_neighbor_distance'] = mash_top_hit[s]['mash_hit_score']
-
-        else:
-            logging.warning("Found sequence: {} without a cluster record, confirm cluster file matches mash file".format(mash_top_hit[s]['top_hit']))
-
-
-
-
-
-    #merge all of the biomarkers if all contig sequences should be treated beloning to one plasmid
-    biomarkers = {}
-
-    if not multi:
-        biomarkers[sample_id] = {'found_replicons':{'acs':[],'types':[]},
-                                 'found_mob':{'acs':[],'types':[]},
-                                 'found_mpf':{'acs':[],'types':[]},
-                                 'found_orit':{'acs':[],'types':[]}}
-
-        tmp = sort_biomarkers(mergeBiomarkers(found_replicons, sample_id))
-        biomarkers[sample_id]['found_replicons']['acs'] = tmp[sample_id]['acs']
-        biomarkers[sample_id]['found_replicons']['types'] = tmp[sample_id]['types']
-
-        tmp = sort_biomarkers(mergeBiomarkers(found_mob, sample_id))
-        biomarkers[sample_id]['found_mob']['acs'] = tmp[sample_id]['acs']
-        biomarkers[sample_id]['found_mob']['types'] = tmp[sample_id]['types']
-
-        tmp = sort_biomarkers(mergeBiomarkers(found_mpf, sample_id))
-        biomarkers[sample_id]['found_mpf']['acs'] =  tmp[sample_id]['acs']
-        biomarkers[sample_id]['found_mpf']['types'] = tmp[sample_id]['types']
-
-        tmp = sort_biomarkers(mergeBiomarkers(found_orit, sample_id))
-        biomarkers[sample_id]['found_orit']['acs'] = tmp[sample_id]['acs']
-        biomarkers[sample_id]['found_orit']['types'] = tmp[sample_id]['types']
-    else:
-        for s in mob_cluster_memberships:
-            if s not in biomarkers:
-                biomarkers[s] = {'found_replicons':{'acs':[],'types':[]},
-                                 'found_mob':{'acs':[],'types':[]},
-                                 'found_mpf':{'acs':[],'types':[]},
-                                 'found_orit':{'acs':[],'types':[]}}
-            if s in found_replicons:
-                biomarkers[s]['found_replicons'] = found_replicons[s]
-            if s in found_mob:
-                biomarkers[s]['found_mob'] = found_mob[s]
-            if s in found_mpf:
-                biomarkers[s]['found_mpf'] = found_mpf[s]
-            if s in found_orit:
-                biomarkers[s]['found_orit'] = found_orit[s]
-
-    biomarkers = predict_mobility(biomarkers)
-
+    mash_input_fasta = fixed_fasta + '.msh'
 
     ncbi = dict_from_alt_key_list(
-        read_file_to_dict(NCBI_PLASMID_TAXONOMY_FILE, NCBI_PLASMID_TAXONOMY_HEADER, separater="\t"), "sample_id")
-    lit = dict_from_alt_key_list(read_file_to_dict(LIT_PLASMID_TAXONOMY_FILE, LIT_PLASMID_TAXONOMY_HEADER, separater="\t"),"sample_id")
-
-    for sid in biomarkers:
-        rep_types = biomarkers[sid]['found_replicons']['types']
-        relaxase_types = biomarkers[sid]['found_mob']['acs']
-        mob_cluster_id = mob_cluster_memberships[sid]['primary_cluster_id']
-        host_range = hostrange(rep_types, relaxase_types, mob_cluster_id,ncbi,lit)
-        for key in host_range:
-            mobtyper_results[sid][key] = host_range[key]
-
-        for key in biomarkers[sid]:
-            if key in mobtyper_results[sid]:
-                mobtyper_results[sid][key] = biomarkers[sid][key]
-
-        if sid in mob_cluster_memberships:
-            for key in mob_cluster_memberships[sid]:
-                if key in mobtyper_results[sid]:
-                    mobtyper_results[sid][key] = mob_cluster_memberships[sid][key]
-
-        if len(biomarkers[sid]['found_replicons']['types']) > 0:
-            mobtyper_results[sid]['rep_type(s)'] = ",".join(biomarkers[sid]['found_replicons']['types'])
-            mobtyper_results[sid]['rep_type_accession(s)'] = ",".join(biomarkers[sid]['found_replicons']['acs'])
-
-        if len(biomarkers[sid]['found_mob']['types']) > 0:
-            mobtyper_results[sid]['relaxase_type(s)'] = ",".join(biomarkers[sid]['found_mob']['types'])
-            mobtyper_results[sid]['relaxase_type_accession(s)'] = ",".join(biomarkers[sid]['found_mob']['acs'])
-
-        if len(biomarkers[sid]['found_mpf']['types']) > 0:
-            mpf_type = determine_mpf_type(biomarkers[sid]['found_mpf']['types'])
-            mobtyper_results[sid]['mpf_type'] = mpf_type
-
-            mobtyper_results[sid]['mpf_type_accession(s)'] = ",".join(biomarkers[sid]['found_mob']['acs'])
-
-        if len(biomarkers[sid]['found_orit']['types']) > 0:
-            mobtyper_results[sid]['orit_type(s)'] = ",".join(biomarkers[sid]['found_orit']['types'])
-            mobtyper_results[sid]['orit_accession(s)'] = ",".join(biomarkers[sid]['found_orit']['acs'])
+        read_file_to_dict(NCBI_PLASMID_TAXONOMY_FILE, NCBI_PLASMID_TAXONOMY_HEADER, separater="\t"),
+        "sample_id")
+    lit = dict_from_alt_key_list(
+        read_file_to_dict(LIT_PLASMID_TAXONOMY_FILE, LIT_PLASMID_TAXONOMY_HEADER, separater="\t"), "sample_id")
 
 
+    if multi:
+        m.mashsketch(input_fasta=fixed_fasta,output_path=mash_input_fasta,sketch_ind=True,num_threads=num_threads)
+        mash_results = parseMash(
+            m.run_mash(reference_db=mash_db, input_fasta=mash_input_fasta, table=False, num_threads=num_threads))
+
+        for seq_id in mash_results:
+            record = {}
+            for field in MOB_TYPER_REPORT_HEADER:
+                if field in contig_info[seq_id]:
+                    record[field] = contig_info[seq_id][field]
+                else:
+                    record[field] = ''
+            record['sample_id'] = seq_id
+            record['num_contigs'] = 1
+            distances = OrderedDict(sorted(mash_results[seq_id].items(), key=itemgetter(1), reverse=False))
+            mash_neighbor_id = next(iter(distances))
+            dist = distances[mash_neighbor_id]
+            record['mash_nearest_neighbor'] = mash_neighbor_id
+            record['mash_neighbor_distance'] = dist
+            record['primary_cluster_id'] = reference_sequence_meta[mash_neighbor_id]['primary_cluster_id']
+            record['secondary_cluster_id'] = reference_sequence_meta[mash_neighbor_id]['secondary_cluster_id']
+            record['mash_neighbor_identification'] = reference_sequence_meta[mash_neighbor_id]['organism']
+            mobtyper_results.append(record)
+
+    else:
+        m.mashsketch(input_fasta=fixed_fasta, output_path=mash_input_fasta, sketch_ind=False, num_threads=num_threads)
+        mash_results = parseMash(m.run_mash(reference_db=mash_db, input_fasta=mash_input_fasta, table=False, num_threads=num_threads))
+        record = {}
+
+        for field in MOB_TYPER_REPORT_HEADER:
+            record[field] = ''
+
+        record['sample_id'] = sample_id
+        fastaSeqStats = calcFastaStats(fixed_fasta)
+        record['md5'] = fastaSeqStats['md5']
+        record['total_length'] = fastaSeqStats['size']
+        record['num_contigs'] = fastaSeqStats['num_seq']
+        record['gc'] = fastaSeqStats['gc_content']
+
+        for seq_id in mash_results:
+            distances = OrderedDict(sorted(mash_results[seq_id].items(), key=itemgetter(1), reverse=False))
+            mash_neighbor_id = next(iter(distances))
+            dist = distances[mash_neighbor_id]
+            record['mash_nearest_neighbor'] = mash_neighbor_id
+            record['mash_neighbor_distance'] = dist
+            record['primary_cluster_id'] = reference_sequence_meta[mash_neighbor_id]['primary_cluster_id']
+            record['secondary_cluster_id'] = reference_sequence_meta[mash_neighbor_id]['secondary_cluster_id']
+            record['mash_neighbor_identification'] = reference_sequence_meta[mash_neighbor_id]['organism']
+
+        record['rep_type(s)'] = []
+        record['rep_type_accession(s)'] = []
+        record['relaxase_type(s)'] = []
+        record['relaxase_type_accession(s)'] = []
+        record['mpf_type'] = []
+        record['mpf_type_accession(s)'] = []
+        record['orit_type(s)'] = []
+        record['orit_accession(s)'] = []
 
 
-    results = []
-    for s in mobtyper_results:
-        results.append(mobtyper_results[s])
 
-    writeReport(results, MOB_TYPER_REPORT_HEADER, report_file)
+        for seq_id in contig_info:
+            record['rep_type(s)'].append(contig_info[seq_id]['rep_type(s)'])
+            record['rep_type_accession(s)'].append(contig_info[seq_id]['rep_type_accession(s)'])
+            record['relaxase_type(s)'].append(contig_info[seq_id]['relaxase_type(s)'])
+            record['relaxase_type_accession(s)'].append(contig_info[seq_id]['relaxase_type_accession(s)'])
+            record['mpf_type'].append(contig_info[seq_id]['mpf_type'])
+            record['mpf_type_accession(s)'].append(contig_info[seq_id]['mpf_type_accession(s)'])
+            record['orit_type(s)'].append(contig_info[seq_id]['orit_type(s)'])
+            record['orit_accession(s)'].append(contig_info[seq_id]['orit_accession(s)'])
+
+        for field in record:
+            tmp = []
+            if record[field] == None :
+                continue
+            if isinstance(record[field],list):
+                length = len(record[field])
+                for i in range(0,length):
+                    tmp += record[field][i].split(',')
+            elif isinstance(record[field],str) and len(record[field]) > 0:
+                tmp += record[field].split(',')
+            if len(tmp) > 0:
+                record[field] = []
+                for d in tmp:
+                    if len(d) > 0:
+                        record[field].append(d)
+
+
+
+        mobtyper_results.append(record)
+
+    for i in range(0,len(mobtyper_results)):
+        record = mobtyper_results[i]
+        bio_markers = sort_biomarkers({0: {'types': record['rep_type(s)'], 'acs': record['rep_type_accession(s)']},
+                         1: {'types': record['relaxase_type(s)'], 'acs': record['relaxase_type_accession(s)']},
+                         2: {'types': record['mpf_type'], 'acs': record['mpf_type_accession(s)']},
+                         3: {'types': record['orit_type(s)'], 'acs': record['orit_accession(s)']},})
+
+        record['rep_type(s)'] = bio_markers[0]['types']
+        record['rep_type_accession(s)'] = bio_markers[0]['acs']
+        record['relaxase_type(s)'] = bio_markers[1]['types']
+        record['relaxase_type_accession(s)'] = bio_markers[1]['acs']
+        record['mpf_type'] = determine_mpf_type(bio_markers[2]['types'])
+        record['mpf_type_accession(s)'] = bio_markers[2]['acs']
+        record['orit_type(s)'] = bio_markers[3]['types']
+        record['orit_accession(s)'] = bio_markers[3]['acs']
+
+
+        if record['mash_neighbor_distance'] <= primary_distance:
+            mob_cluster_id = record['primary_cluster_id']
+        else:
+            mob_cluster_id = None
+        host_range = hostrange(record['rep_type(s)'], record['relaxase_type_accession(s)'], mob_cluster_id, ncbi, lit)
+
+        for field in host_range:
+            record[field] = host_range[field]
+
+
+        for field in record:
+            if isinstance(record[field],list):
+                record[field] = ",".join(record[field])
+        mobtyper_results[i] = record
+
+
+    writeReport(mobtyper_results, MOB_TYPER_REPORT_HEADER, report_file)
